pyproject.toml
[project]
name = "mlx-edges"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mlx>=0.29.1",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
]
